Fix long-time bug with ajv/window-config in server vs. normal Emacs
If this was set to `after-init`, then the daemon would hangup and would
not work properly. It was probably waiting for ajv/window-config to finish
running, but that wouldn't since there is no window yet.

At the same time, without a hook, normal Emacs wouldn't show the preferred
window config as setup. The fix was to use the `emacs-startup` hook instead.